Remove Tile.cs, create TileController.cs
From what I can tell, unity scripts tend to be called BlahController, while the object itself will be called Blah. For our use case, I have changed Tile.cs to TileController.cs and also made the script extend from MonoBehaviour - this is because scripts need to extend from MonoBehaviour in order to be attached to a GameObject. Currently, TileController has a method for resizing the tile, as the location of the tile can be set when Instantiating it within the GridController